This adds play and pause event handling to the mpris providers
Adding the play and pause events in addition to the play-pause
event. This makes it easier to pause the player externally if you
don't have access to the current status.

I am using an instance variable, because this.player.playbackStatus
doesn't seem to be available all of the time.

This fixes #440